Fix GSI ncdiag converter dimension issue (#50)
Fixes using a single dictionary that gets modified for all GSI ncdiags
to a separate dictionary for each instance of the class. This solves the
issue of Channel dimensions showing up for conventional data after
radiance data has been processed in a loop.

Also changed `.nc4` to `.nc` to match other filename conventions in
other components