da-utils companion to GW obs staging and analysis stats refactor (#49)
This PR, a companion to
https://github.com/NOAA-EMC/global-workflow/pull/4306 and
https://github.com/NOAA-EMC/GDASApp/pull/1999 makes a small change to
the obs statistics application, so that the key `obs spaces` becomes
`observers`. This way, the YAML structure for obs spaces is identical to
all other JEDI configuration YAMLs and the `clean_empty_obsspaces()`
method in the GW `Jedi` class can be applied to the configuration YAML
for this application as well.

---------

Co-authored-by: Cory Martin <cory.r.martin@noaa.gov>